backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
requests==2.31.0

# LLM and Embeddings
openai==1.3.9
langchain==0.1.1
langchain-openai==0.0.7
sentence-transformers==2.2.2

# Vector Database
qdrant-client==2.7.0

# Document Processing
pypdf==3.17.1
pymupdf==1.23.8
pytesseract==0.3.10
pillow==10.1.0

# Language Detection
langdetect==1.0.9

# Data Processing
numpy==1.26.3
pandas==2.1.3

# Caching
cachetools==5.3.2

# CORS
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
